+++ frontend/src/components/models/task/ChildrenTasksDetails.tsx
@@ -8,40 +8,41 @@ import {
   Typography,
 } from "@mui/material";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
-import { DivisionIncludeUserAvatar, ParentTask } from "../../../types";
+import { ChildTask } from "../../../types";
 import { TaskContentTypography } from "./TaskContentTypography";
 import { DatetimeFormat } from "../../ui/DatetimeFormat";
 import { PriorityStack } from "./PriorityStack";
 import { TaskCardAvatar } from "./TaskCardAvatar";
 
 type Props = {
-  division: DivisionIncludeUserAvatar;
-  parentTask: ParentTask;
+  childrenTasks: ChildTask[];
 };
 
-export const ParentTasksCard = (props: Props) => {
-  const { division, parentTask } = props;
+export const ChildrenTasksDetails = (props: Props) => {
+  const { childrenTasks } = props;
 
   return (
-    <Card sx={{ width: 550, p: 2 }}>
-      <CardHeader title="親タスク情報" />
-      <CardContent>
-        <Divider sx={{ mb: 2 }} />
-        <Grid2 container rowSpacing={1} p={0}>
+    <Grid2 container rowSpacing={1}>
+      {childrenTasks?.map((childTask) => (
+        <div key={childTask.id}>
           <Grid2 xs={12}>
             <Typography color="text.secondary" variant="body2" component="div">
-              {DatetimeFormat(division.created_at)}
+              {DatetimeFormat(childTask.division.created_at)}
             </Typography>
           </Grid2>
           <Grid2 xs={12}>
             <Stack direction="row" spacing={1} alignItems="center">
-              <TaskCardAvatar avatar={division.avatar} />
-              <Typography>{division.user.name}から分担されました。</Typography>
+              <TaskCardAvatar avatar={childTask.division_avatar} />
+              <Typography>
+                {childTask.division.user.name}が分担を作成しました。
+              </Typography>
             </Stack>
           </Grid2>
-          {division.comment ? (
+          {childTask.division.comment ? (
             <Grid2 xs={12}>
-              <Typography>コメント：{division.comment}</Typography>
+              <Typography mt={1}>
+                コメント：{childTask.division.comment}
+              </Typography>
             </Grid2>
           ) : null}
           <Grid2 xs={12}>
@@ -50,59 +51,60 @@ export const ParentTasksCard = (props: Props) => {
                 title={
                   <Stack direction="column" spacing={1}>
                     <Stack direction="row" spacing={1} alignItems="center">
-                      <TaskCardAvatar avatar={parentTask.avatar} />
+                      <TaskCardAvatar avatar={childTask.avatar} />
                       <Typography variant="body1" component="div">
-                        {parentTask.user.name}のタスク
+                        {childTask.user.name}のタスク
                       </Typography>
                     </Stack>
                     <Typography
                       variant="h6"
                       component={Link}
-                      to={`/tasks/${parentTask.id}`}
+                      to={`/tasks/${childTask.id}`}
                       sx={{ color: "inherit" }}
                     >
-                      {parentTask.title}
+                      {childTask.title}
                     </Typography>
                   </Stack>
                 }
               />
               <CardContent>
                 <Grid2 container rowSpacing={2} p={0}>
-                  {parentTask.description ? (
+                  {childTask.description ? (
                     <Grid2 xs={12}>
                       <TaskContentTypography
                         subtitle="詳細"
-                        body={parentTask.description}
+                        body={childTask.description}
                       />
                     </Grid2>
                   ) : null}
                   <Grid2 xs={6}>
-                    <PriorityStack value={parentTask.priority} />
+                    <PriorityStack value={childTask.priority} />
                   </Grid2>
                   <Grid2 xs={6}>
                     <TaskContentTypography
                       subtitle="納期"
-                      body={DatetimeFormat(parentTask.deadline)}
+                      body={DatetimeFormat(childTask.deadline)}
                     />
                   </Grid2>
                   <Grid2 xs={6}>
                     <TaskContentTypography
                       subtitle="ステータス"
-                      body={parentTask.is_done ? "完了済み" : "未了"}
+                      body={childTask.is_done ? "完了済み" : "未了"}
                     />
                   </Grid2>
                   <Grid2 xs={6}>
                     <TaskContentTypography
                       subtitle="進捗率"
-                      body={`${parentTask.rate_of_progress}%`}
+                      body={`${childTask.rate_of_progress}%`}
                     />
                   </Grid2>
                 </Grid2>
               </CardContent>
             </Card>
           </Grid2>
-        </Grid2>
-      </CardContent>
-    </Card>
+          <Divider sx={{ my: 4 }} />
+        </div>
+      ))}
+    </Grid2>
   );
 };
